added custom hook for generating column definitions for SubmissionDataGrid
